backend-requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-mistralai
chromadb
sentence-transformers
python-dotenv
requests